#1250 Moved meta-propety related test cases for persistent entities into a more appropriate package.
