+++ platform-dao/src/test/java/ua/com/fielden/platform/entity/meta/MetaPropertyFromPersistancePerspectiveTest.java
@@ -1,4 +1,4 @@
-package ua.com.fielden.platform.entity;
+package ua.com.fielden.platform.entity.meta;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
